Multi Window MVP - Base
Requires modifications to the Phosphor library to not throw errors when
widgets are attached to elements not contained in the original DOM.
Make sure to read the README and apply the changes to Phosphor accordingly.

Primary changes:
- Add concept of extractable widgets. Only widgets implementing the interface can be extracted
- Add WidgetExtractionHandler that encapsulates logic to move widgets to new windows.
- Only plugin/vscode extension views can be extracted
- Allow opening "native" windows in Electron
- Bridge plugin communication
- Add extracted CSS and CatCoding VS Code extension

Secure messaging between external webview and Theia:
Webviews only accept messages from its direct parent window. This is necessary to avoid Cross Site Scripting (XSS).
To make the messaging work in external windows, messages are sent to the external widget and then delegated to the webview's iframe.
Thereby, the external window only accepts messages from its opener which is the theia main window.
To achieve this, a webview knows the external window its is in (if any). It then sends messages to this window instead of directly to the webview iframe.